+++ src/main/scala/io/viash/ViashTest.scala
@@ -70,10 +70,15 @@ object ViashTest extends Logging {
     parentTempPath: Option[Path] = None, 
     cpus: Option[Int], 
     memory: Option[String],
-    dryRun: Option[Boolean] = None
+    dryRun: Option[Boolean] = None,
+    deterministicWorkingDirectory: Option[String] = None,
   ): ManyTestOutput = {
     // create temporary directory
-    val dir = IO.makeTemp("viash_test_" + configNameLens.get(appliedConfig), parentTempPath)
+    val dir = IO.makeTemp(
+      name = deterministicWorkingDirectory.getOrElse(s"viash_test_${configNameLens.get(appliedConfig)}_"),
+      parentTempPath = parentTempPath,
+      addRandomized = deterministicWorkingDirectory.isEmpty
+    )
     if (!quiet) infoOut(s"Running tests in temporary directory: '$dir'")
 
     DependencyResolver.createBuildYaml(dir.toString())