+++ src/main/scala/io/viash/config/dependencies/Repository.scala
@@ -33,7 +33,7 @@ import java.nio.file.{Path, Paths, Files}
   """repositories:
     |  - name: openpipelines-bio
     |    type: github
-    |    uri: openpipelines-bio/modules
+    |    repo: openpipelines-bio/modules
     |    tag: 0.3.0
     |""".stripMargin,
     "yaml")