+++ src/test/scala/io/viash/e2e/ns_test/MainNSTestNativeSuite.scala
@@ -227,6 +227,37 @@ class MainNSTestNativeSuite extends AnyFunSuite with BeforeAndAfterAll {
     }
   }
 
+  test("Check namespace test output with deterministic working directory") {
+    // create a new unique temporary folder
+    val temporaryFolder = IO.makeTemp(s"viash_${this.getClass.getName}_")
+    val tempFolStr = temporaryFolder.getFileName().toString()
+
+    val testOutput = TestHelper.testMain(
+      "ns", "test",
+      "--src", nsPath,
+      "--deterministic_working_directory", tempFolStr
+    )
+
+    assert(testOutput.stderr.contains(s"The working directory for the namespace tests is ${temporaryFolder}"))
+
+    // Test inclusion of a header
+    val regexHeader = s"^\\s*${ViashNamespace.columnHeaders.mkString("\\s*")}".r
+    assert(regexHeader.findFirstIn(testOutput.stdout).isDefined, s"\nRegex: ${regexHeader.toString}; text: \n${testOutput.stdout}")
+
+    for (
+      (component, steps) <- components;
+      (step, resultPattern) <- steps
+    ) {
+      val regex = s"""testns\\s*$component\\s*executable\\s*native\\s*$step$resultPattern""".r
+      assert(regex.findFirstIn(testOutput.stdout).isDefined, s"\nRegex: '${regex.toString}'; text: \n${testOutput.stdout}")
+    }
+
+    val regexBuildError = raw"Reading file \'.*/src/ns_error/config\.vsh\.yaml\' failed".r
+    assert(regexBuildError.findFirstIn(testOutput.stderr).isDefined, "Expecting to get an error because of an invalid yaml in ns_error")
+
+    assert(testOutput.stderr.contains("The status of the component 'ns_power' is set to deprecated."))
+  }
+
   override def afterAll(): Unit = {
     IO.deleteRecursively(temporaryFolder)
   }