+++ CHANGELOG.md
@@ -95,9 +95,13 @@ Additionally changes are made to sanitize the built config output and include ad
 
 * `docker setup strategy`: Fix inconsistencies in the documentation (PR #657).
 
+* `repositories`: Fix `uri` -> `repo` in the repositories documentation (PR #682).
+
 ## NEW FUNCTIONALITY
 
 * `viash test` and `viash ns test`: Add a hidden `--dry_run` option to build the tests without executing them (PR #676).
+
+* `viash test` and `viash ns test`: Add a hidden `--deterministic_working directory` argument to use a fixed directory path (PR #683).
 
 ## BUG FIXES